requirements.txt
click==8.1.3
Flask==2.2.2
importlib-metadata==6.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
paho-mqtt==1.6.1
psycopg2==2.9.5
waitress==2.1.2
Werkzeug==2.2.2
zipp==3.11.0
